update verification routes
