+++ src/routes/account/verification/+page.js
@@ -1,0 +1,8 @@
+import { redirect } from "@sveltejs/kit";
+
+export async function load({ parent }) {
+	const { user } = await parent();
+	if (user) {
+		throw redirect(307, "/account");
+	}
+}
