+++ src/routes/account/verification/+page.svelte
@@ -1,0 +1,18 @@
+<script>
+	import { page } from "$app/stores";
+
+	const token = new URLSearchParams($page.url.hash.substring(1)).get("access_token");
+
+	if (token) {
+		setCookie();
+	}
+
+	async function setCookie() {
+		fetch("/account/verification", {
+			method: "POST",
+			body: JSON.stringify({ token })
+		});
+	}
+</script>
+
+<h1>Waiting for verification...</h1>
